requirements.txt
awscli==1.16.144
awsebcli==3.15.0
blessed==1.15.0
boto3==1.9.134
botocore==1.12.134
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
decorator==4.4.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.0.2
future==0.16.0
idna==2.7
imageio==2.5.0
imageio-ffmpeg==0.3.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
moviepy==1.0.0
multidict==4.5.2
nltk==3.4.1
numpy==1.16.3
pandas==0.24.2
pathspec==0.5.9
Pillow==6.0.0
proglog==0.1.9
pyasn1==0.4.5
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
semantic-version==2.5.0
six==1.11.0
sklearn==0.0
termcolor==1.1.0
texttable==0.9.1
tqdm==4.31.1
urllib3==1.24.2
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.2
wordcloud==1.5.0
youtube-dl==2019.4.17
